partI Equator passed

+++ ft_strlen.c
@@ -10,9 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned long int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
-	int counter;
+	size_t counter;
 
 	counter = 0;
 	while (*s)
